requirements.txt
praw
google.cloud
google.cloud.speech
google-cloud-storage